+++ vss_cli/const.py
@@ -5,7 +5,7 @@ import pkg_resources
 
 PACKAGE_NAME = "vss_cli"
 
-__version__ = "0.7.1"
+__version__ = "0.8.0"
 
 REQUIRED_PYTHON_VER = (3, 6, 4)
 
@@ -272,11 +272,7 @@ COLUMNS_VM_GUEST = [
     ("guest_id", "os.guest_id"),
     ("running_status", "tools.running_status"),
 ]
-COLUMNS_VM_GUEST_OS = [
-    ("guest_full_name",),
-    ("guest_id",),
-    ("guest_family",),
-]
+COLUMNS_VM_GUEST_OS = [("guest_full_name",), ("guest_id",), ("guest_family",)]
 COLUMNS_VM_GUEST_IP = [
     ("ip_address", "ip_address"),
     ("mac_address", "mac_address"),
@@ -394,14 +390,18 @@ COLUMNS_VM_CONTROLLERS = [('scsi.count',)]
 COLUMNS_EXTRA_CONFIG = [("key",), ("value",)]
 COLUMNS_VSS_OPTIONS = [("options", "[*]")]
 COLUMNS_GROUP = [
-    ("cn",),
+    ("name",),
     ("description",),
-    ("create_timestamp",),
-    ("modify_timestamp",),
-    ("unique_member_count",),
-    ("unique_member", "unique_member[*].uid"),
+    ("members.count",),
+    ("ldap.last_sync",),
 ]
-COLUMNS_GROUPS = [("groups", "groups[*]")]
+COLUMNS_GROUPS = [("id",), ("name",), ("description",), ('members.count',)]
+COLUMNS_GROUP_MEMBERS = [
+    ("username",),
+    ("first_name",),
+    ("last_name",),
+    ("email",),
+]
 COLUMNS_ROLE = [
     ("name",),
     ("description",),